Refactor/move singleton out of config module (#3161)
