+++ .gitignore
@@ -20,6 +20,7 @@ log-ingestion.txt
 logs
 *.log
 *.mp3
+mem.sqlite3
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
