+++ autogpt/config/__init__.py
@@ -3,12 +3,9 @@ This module contains the configuration classes for AutoGPT.
 """
 from autogpt.config.ai_config import AIConfig
 from autogpt.config.config import Config, check_openai_api_key
-from autogpt.config.singleton import AbstractSingleton, Singleton
 
 __all__ = [
     "check_openai_api_key",
-    "AbstractSingleton",
     "AIConfig",
     "Config",
-    "Singleton",
 ]
