+++ autogpt/config/config.py
@@ -8,7 +8,7 @@ from auto_gpt_plugin_template import AutoGPTPluginTemplate
 from colorama import Fore
 from dotenv import load_dotenv
 
-from autogpt.config.singleton import Singleton
+from autogpt.singleton import Singleton
 
 load_dotenv(verbose=True, override=True)
