+++ autogpt/memory/base.py
@@ -1,7 +1,8 @@
 """Base class for memory providers."""
 import abc
 
-from autogpt.config import AbstractSingleton, Config
+from autogpt.config import Config
+from autogpt.singleton import AbstractSingleton
 
 cfg = Config()
