+++ autogpt/singleton.py
